+++ src/shell/cmds/cmd_vmstat.c
@@ -14,10 +14,10 @@ void cmd_vmstat(const char *args) {
 
     kprintf("Virtual Memory Statistics:\n");
     kprintf("-------------------------\n");
-    kprintf("Total RAM:      %u KB (%u pages)\n", (unsigned long)total_kb, (unsigned long)stats.total_pages);
-    kprintf("Used RAM:       %u KB (%u pages)\n", (unsigned long)used_kb, (unsigned long)stats.used_pages);
-    kprintf("Free RAM:       %u KB (%u pages)\n", (unsigned long)free_kb, (unsigned long)stats.free_pages);
+    kprintf("Total RAM:      %u KB (%u pages)\n", total_kb, stats.total_pages);
+    kprintf("Used RAM:       %u KB (%u pages)\n", used_kb, stats.used_pages);
+    kprintf("Free RAM:       %u KB (%u pages)\n", free_kb, stats.free_pages);
     kprintf("\n");
-    kprintf("Uptime ticks:   %u\n", (unsigned long)libc_uptime_ticks());
-    kprintf("Active PIDs:    %u\n", (unsigned long)libc_getpid());
+    kprintf("Uptime ticks:   %llu\n", (unsigned long long)libc_uptime_ticks());
+    kprintf("Active PIDs:    %llu\n", (unsigned long long)libc_getpid());
 }
